Généralisation de Increment -> Double et documentation

+++ src/main/java/net/gunivers/listgenerator/functionality/increment/Increment.java
@@ -5,15 +5,25 @@ import java.util.ArrayList;
 import net.gunivers.listgenerator.functionality.Functionality;
 import net.gunivers.listgenerator.util.Call;
 
+/**
+ * @author Oromis
+ * A function to increment at each loop
+ */
 public class Increment extends Functionality {
 
+	/**
+	 * @param initValue first value of the loop
+	 * @param incr number to add at each loop
+	 * @param nbLoop number of loop
+	 * @return an ArrayList<String> with all the value replacing the tag
+	 */
 	@Call
-	public ArrayList<String> incremente(Integer initValue, Integer incr, Integer nbLoop) {
+	public ArrayList<String> incremente(Double initValue, Double incr, Integer nbLoop) {
 		ArrayList<String> save = new ArrayList<String>();
-		save.add(Integer.toString(initValue));
+		save.add(Double.toString(initValue));
 		
 		for(int i = 0; i < nbLoop; i++)
-			save.add(Integer.toString(initValue += incr));
+			save.add(Double.toString(initValue += incr));
 		
 		return save;
 	}
